progress bar setup
