+++ src/index.js
@@ -1,26 +1,11 @@
-import React, { Component } from "react";
-import { BrowserRouter, Switch, Route } from "react-router-dom";
-import ReactDOM from "react-dom";
-import "./index.css";
-import "bootstrap/dist/css/bootstrap.min.css";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
+import React from "react";
+import {App} from "./App";
+import {Provider} from "react-redux";
 
-class Root extends Component {
-  render() {
-    return (
-      <BrowserRouter>
-        <Switch>
-          <Route exact path="/" component={App} />
-        </Switch>
-      </BrowserRouter>
-    );
-  }
-}
 
-ReactDOM.render(<Root />, document.getElementById("root"));
+dotenv.config();
 
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: http://bit.ly/CRA-PWA
-serviceWorker.unregister();
+<App/>
+
+
+ReactDOM.render(<App />, document.getElementById("root"));
