+++ src/router.js
@@ -1,0 +1,27 @@
+import React, { Component } from "react";
+import { BrowserRouter, Switch, Route } from "react-router-dom";
+import ReactDOM from "react-dom";
+import "./index.css";
+import "bootstrap/dist/css/bootstrap.min.css";
+import App from "./App";
+import * as serviceWorker from "./serviceWorker";
+
+
+class Root extends Component {
+  render() {
+    return (
+      <BrowserRouter>
+        <Switch>
+          <Route exact path="/" component={App} />
+        </Switch>
+      </BrowserRouter>
+    );
+  }
+}
+
+ReactDOM.render(<Root />, document.getElementById("root"));
+
+// If you want your app to work offline and load faster, you can change
+// unregister() to register() below. Note this comes with some pitfalls.
+// Learn more about service workers: http://bit.ly/CRA-PWA
+serviceWorker.unregister();
